# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AiMessageRole = typeof AiMessageRole[keyof typeof AiMessageRole];


export const AiMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AiMessage {
  role: AiMessageRole;
  content: string;
}

export type AiToolInputSchema = { [key: string]: unknown };

export interface AiTool {
  name: string;
  description?: string;
  input_schema: AiToolInputSchema;
}

export interface AiMessageInput {
  model?: string;
  max_tokens: number;
  /** @nullable */
  system?: string | null;
  messages: AiMessage[];
  /** @nullable */
  tools?: AiTool[] | null;
}

export type AiMessageResponseContentItem = { [key: string]: unknown };

export type AiMessageResponseUsage = { [key: string]: unknown };

export interface AiMessageResponse {
  id: string;
  type: string;
  role: string;
  content: AiMessageResponseContentItem[];
  model: string;
  /** @nullable */
  stop_reason?: string | null;
  usage?: AiMessageResponseUsage;
}

export interface OmExtractInput {
  /** The raw OM text extracted from PDF */
  text: string;
}

/**
 * The extracted deal data matching the OM schema
 */
export type OmExtractResponseData = { [key: string]: unknown };

export interface OmExtractResponse {
  /** The extracted deal data matching the OM schema */
  data: OmExtractResponseData;
  /** Whether the full tenant list was extracted without truncation */
  tenantsComplete: boolean;
}

export interface ApiError {
  error: string;
}