# lib/api-zod/src/generated/types/aiMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiMessageRole } from './aiMessageRole';

export interface AiMessage {
  role: AiMessageRole;
  content: string;
}