# lib/api-zod/src/generated/types/aiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiMessage } from './aiMessage';
import type { AiTool } from './aiTool';

export interface AiMessageInput {
  model?: string;
  max_tokens: number;
  /** @nullable */
  system?: string | null;
  messages: AiMessage[];
  /** @nullable */
  tools?: AiTool[] | null;
}