# lib/api-zod/src/generated/types/aiMessageResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiMessageResponseContentItem } from './aiMessageResponseContentItem';
import type { AiMessageResponseUsage } from './aiMessageResponseUsage';

export interface AiMessageResponse {
  id: string;
  type: string;
  role: string;
  content: AiMessageResponseContentItem[];
  model: string;
  /** @nullable */
  stop_reason?: string | null;
  usage?: AiMessageResponseUsage;
}