# lib/api-zod/src/generated/types/aiTool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiToolInputSchema } from './aiToolInputSchema';

export interface AiTool {
  name: string;
  description?: string;
  input_schema: AiToolInputSchema;
}