# lib/api-zod/src/generated/types/omExtractResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OmExtractResponseData } from './omExtractResponseData';

export interface OmExtractResponse {
  /** The extracted deal data matching the OM schema */
  data: OmExtractResponseData;
  /** Whether the full tenant list was extracted without truncation */
  tenantsComplete: boolean;
}